requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
openai==1.12.0
google-cloud-speech==2.25.1
google-cloud-texttospeech==2.15.1
sounddevice==0.4.6
soundfile==0.12.1
numpy==1.26.4
reportlab==4.1.0
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
requests==2.31.0
python-multipart==0.0.9
asgiref==3.7.2
whisper==1.1.10
markdown==3.3.4
textblob==0.15.3
weasyprint==52.5
ffmpeg-python==0.2.0
SpeechRecognition==3.10.0